Refactor plugin ids (#299)
_This is PR 1 of 3. The other two are: #300 and #301_

This PR is a segue to the next PR which does the actual job.

Unfortunately, we cannot apply Java plugin if the user applies the
Android plugin because these two conflict on each other. We trust the
user to apply the correct plugin in their Gradle script but we wouldn't
force it.

To avoid the performance overhead when checking for available plugins,
we should reference them using their plugin IDs instead of actual class.
Therefore the type of collection was changed from `<Class<out
Plugin<Project>>>` to just `<String>`.

Because this change only alters the internals but does not alter the
behavior, i.e. "everything works as before", I call it a "refactoring
PR". To get the idea how this change is used, check out the next PR #300